--- src/main/java/chess/SparkChessApplication.java
@@ -1,10 +1,0 @@
-package chess;
-
-//import chess.controller.WebChessController;
-//
-//public class SparkChessApplication {
-//    public static void main(String[] args) {
-//        WebChessController webChessController = new WebChessController();
-//        webChessController.run();
-//    }
-//}
