--- src/main/java/chess/WebApplication.java
@@ -1,12 +1,0 @@
-//package chess;
-//
-//import chess.controller.WebChessController;
-//
-//public class WebApplication {
-//
-//    public static void main(String[] args) {
-//        WebChessController webChessController = new WebChessController();
-//        webChessController.run();
-//    }
-//
-//}
